+++ Sources/RedBlackTreeModule/RedBlackTreeMultiSet.swift
@@ -966,13 +966,6 @@ extension RedBlackTreeMultiSet {
   where Element: Sendable {}
 #endif
 
-#if false
-#if swift(>=5.5)
-  extension RedBlackTreeMultiSet.SubSequence: @unchecked Sendable
-  where Element: Sendable {}
-#endif
-#endif
-
 // MARK: - Init naive
 
 extension RedBlackTreeMultiSet {