+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.2.1] - 2025-?-?
+### Added
+- サブシーケンスのサブスクリプトにRangeExpression対応を追加
 
 ## [0.2.0] - 2025-10-12
 ### Added
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RedBlackTreeSliceを追加
 - RangeExpression対応を追加
 ### Changed
+- RedBlackTreeMultiMapのKeyValueをタプルから構造体に変更
 - IndexをRedBlackTreeIndexに変更
 - SubSequenceをRedBlackTreeSliceに変更
 - 内部Indexの比較アルゴリズムを変更